% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_generate_params.R
\name{ABC_params}
\alias{ABC_params}
\title{Simulate parameters and summaries from a model.}
\usage{
ABC_params(
  sample_num,
  point_pattern_target = realepmild,
  point_pattern_bp_target = realbpMILD,
  normal_patterns = realep,
  normal_bp_patterns = realbp,
  nsim = 100,
  epsilon = 5,
  simulate_healthy = FALSE
)
}
\arguments{
\item{{sample_num}}{The sample number}

\item{{normal_patterns}}{A  list of two-dimensional point pattern for the end points to be thinned(object of class "ppp").
Should have a mark "Tree".}

\item{{normal_bp_patterns}}{A list of two-dimensional point patterns for the base points to be thinned(object of class "ppp").
Should have a mark "Tree".}

\item{{point_pattern_target}}{A list of two-dimensional point patterns for the end points
(object of class "ppp"). The number of point in this pattern give the target end point in the thinned pattern.
Should have a mark "Tree".}

\item{{point_pattern_bp_target}}{A list of two-dimensional point patterns for the base points to be thinned(object of class "ppp").
The number of points in this pattern give the target base point in the thinned pattern.
Should have a mark "Tree".}

\item{{weight_bp}}{A scaling parameter for the probability function}

\item{{simulate_healthy}}{A variable deciding whether to simulate or use the empirical healthy data as a starting point}

\item{{epsilon}}{Parameter controlling what healthy point patterns to be thinned.
The rule is that if a pattern has at least target_bp+ epsilon base points then it will be thinned.}

\item{{nsim}}{Number of simulations
@return A list with columns P1 and S1.}
}
\description{
Creates param and sumstat vector for an abc rejection algorithm.
The parameter is taken from a Uniform(0,0.2) dostrobution and S1 is calculated as
the radius r such that the OVERALL empty space function F(r)=0.5.
}
\examples{
ABC_SAMP = ABC_params(nsim=15000,sample_num = 1)
}
