% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cluster_marks.R
\name{add_cluster_size_marks}
\alias{add_cluster_size_marks}
\title{Adds a cluster size mark to a point pattern.}
\usage{
add_cluster_size_marks(point_pattern)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
A Two-dimensional marked point pattern(object of class "ppp") with mark "size" (the number of points in each cluster)
}
\description{
Adds a size (cluster size) mark to each point in the pattern. This function uses the get_cluster_size() function to determine the cluster sizes.
}
\examples{
add_cluster_size_marks(realep[[1]])
}
