% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_convex_hull.R
\name{area_convex_hull}
\alias{area_convex_hull}
\title{Areas of reactive territories}
\usage{
area_convex_hull(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
The areas of reactive territories (the convex hull of the nerve trees including both end and base points)
}
\description{
Computes the areas of reactive territories given two point patterns( base and end point patterns)
}
\examples{
area_convex_hull(realep[[1]], realbp[[1]])
}
