% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_random_marks.R
\name{assign_random_marks}
\alias{assign_random_marks}
\title{Assigns random marks to a pattern}
\usage{
assign_random_marks(point_pattern)
}
\arguments{
\item{{point_pattern}}{Two-dimensional point pattern (object of class "ppp").}
}
\value{
Two-dimensional marked point pattern (object of class "ppp"). The marks "point_id" will be added to the pattern.
}
\description{
This function assigns random marks to a pattern. Mainly used in the Ind_thin() function to randomly thin a pattern to a fixed number of points
}
\examples{
assign_random_marks(realep[[1]])
}
