% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size_histogram.R
\name{create_cluster_size_histogram}
\alias{create_cluster_size_histogram}
\title{Creates a histogram of the cluster sizes.}
\usage{
create_cluster_size_histogram(
  point_pattern_mild = realepmild,
  point_pattern = realep,
  point_pattern_thinned,
  groupnames = c("Healthy", "Mild", "Thinned")
)
}
\arguments{
\item{{point_pattern}}{A list of two-dimensional marked point patterns (objects of class "ppp").
Should contain a mark "Tree". Default = realep (the list containing the healthy end point patterns)}

\item{{point_pattern_mild}}{A list of two-dimensional marked point patterns (objects of class "ppp").
Should contain a mark "Tree". Default = realepmild (the list containing the mild diabetic end point patterns)}

\item{{point_pattern_thinned}}{A list of two-dimensional marked point patterns (objects of class "ppp").
Should contain a mark "Tree". This list should contain the end point patterns obtained after thinning}

\item{{groupnames}}{An argument defining the groups that are compared. The default value is groupnames = c("Healthy","Mild","Thinned")}
}
\value{
Plots a histogram
}
\description{
Creates a histogram comparing the cluster sizes between patterns in three groups.
}
\examples{
create_cluster_size_histogram(point_pattern_thinned = X_thin_end)
}
