% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_convex_hull_plots.R
\name{create_convex_hull_plot}
\alias{create_convex_hull_plot}
\title{Plots reactive territories}
\usage{
create_convex_hull_plot(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
Plots the reactive territories (the convex hull of the nerve trees including both end and base points)
}
\description{
PLOTS THE CONVEX HULL OF THE BASE AND END POINT PATTERNS
}
\examples{
create_convex_hull_plot(realep[[1]],realbp[[1]])
}
