% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_density_plots.R
\name{create_density_plot}
\alias{create_density_plot}
\title{Prior/posterior predictive plots}
\usage{
create_density_plot(V1, V2, observed, xwin = c(30, 100), ywin = c(0, 80))
}
\arguments{
\item{{V1}}{A numerical vector for summary statistic S1}

\item{{V2}}{A numerical vector for summary statistic S2}

\item{{observed}}{The observed c(S1,S2) summary statistic}

\item{{xwin}}{Limits of the x-dimension of the observational window}

\item{{ywin}}{Limits of the y-dimension of the observational window}
}
\value{
A density plot
}
\description{
Creates Prior/posterior predictive plots given vectors containing summary statistics S1 and S2
obtained from parameters from the prior or posterior
}
\examples{
create_density_plot (V1 = ABC_summary_1,V2=ABC_summary_2,observed=c(53.49,25.29))
}
