% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpp3_from_df.R
\name{create_pp3_from_df}
\alias{create_pp3_from_df}
\title{Create a pp3 object}
\usage{
create_pp3_from_df(df, type, use.marks = TRUE)
}
\arguments{
\item{{df}}{A dataframe with columns
\itemize{
\item $X : X coordinate
\item $Y : Y coordinate
\item $Z : Z coordinate
\item $Tree : An integer denoting the tree id.
}}

\item{{type}}{The type of point. If type = 1 then it is a base point and if type = 0 then it is an end point.}

\item{{use.marks}}{Whether to add marks to the pattern or not. Defauls=TRUE.}
}
\value{
A three-dimensional point pattern (object of class "pp3") with marks Tree and type.
}
\description{
Creates a three dimensional marked point pattern with marks "Tree" and "type" from a dataframe.
}
\examples{
create_pp3_from_df(realbpdf[[1]],1)
}
