% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xsim.R
\name{create_xsim}
\alias{create_xsim}
\title{Create an xsim object}
\usage{
create_xsim(X)
}
\arguments{
\item{{X}}{Three-dimensional point pattern (object of class "pp3").}
}
\value{
A list with $x~coordinates $bbox~bounding box
}
\description{
Creates a list with $x~coordinates $bbox~bounding box to be used in the Kest_directional function
}
\examples{
create_xsim(rpoispp3(10))
}
