% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_thinning_tree.R
\name{dependent_thinning}
\alias{dependent_thinning}
\title{Function for tree-wise thinning}
\usage{
dependent_thinning(point_pattern, P)
}
\arguments{
\item{{point_pattern}}{A two-dimensional point pattern(object of class "ppp" in spatstat).
Should have a mark "Tree".}

\item{{P}}{A vector of probabilities to retain each tree. We can get such a vector with the THINING_PROPABILITY_() function.}
}
\value{
A thinned point pattern
}
\description{
This function will thin trees(all points in the tree are removed) according to the vector of probabilities
. The main difficulty is to find a model for the probabilities,
i.e what nerve trees should be more probable to be removed. For instance, we can use PROBABILITY_THINNING_()
to create such a vector.
}
\examples{
dependent_thinning( marked_point_pattern , probabilities)
}
