% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_thinning_ep.R
\name{dependent_thinning_ep}
\alias{dependent_thinning_ep}
\title{Function for point-wise thinning}
\usage{
dependent_thinning_ep(point_pattern, P)
}
\arguments{
\item{{point_pattern}}{A two-dimensional point pattern(object of class "ppp" in spatstat)}

\item{{P}}{A vector of probabilities. We can get such a vector with the THINING_PROPABILITY_() function.}
}
\value{
A thinned point pattern
}
\description{
This function will thin every point according to the vector of probabilities
. The main difficulty is to find a model for the probabilities,
i.e what points should be more probable to be removed. For instance, we can use PROBABILITY_THINNING_()
to create such a vector.
}
\examples{
dependent_thinning_ep( point_pattern , probabilities)
}
