% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_regrowth.R
\name{endpoint_regrowth}
\alias{endpoint_regrowth}
\title{Regrowth model}
\usage{
endpoint_regrowth(
  point_pattern,
  point_pattern_bp,
  num_ep,
  hc = 1,
  R = 20,
  weight_ep = 0.045,
  scheme = "dependent",
  Use.reactives = "FALSE"
)
}
\arguments{
\item{{point_pattern}}{A two-dimensional point pattern (object of class "ppp").
Should have a mark "Tree.}

\item{{point_pattern_bp}A}{two-dimensional point pattern (object of class "ppp").
Should have a mark "Tree.}

\item{{num_ep}}{Target number of end points}

\item{{hc}}{Hardcore parameter}

\item{{R}}{Radius parameter in the regrowth model when adding points}

\item{{weight_ep}}{Scaling parameter in the probabiity function.}

\item{{scheme}}{The thinning scheme for the end points. Either "dependent" or "random". Default is "dependent"}

\item{{Use.reactives}}{A parameter determining how the regrowth model should propose new points.
If true the point is suggested in the reactive territory else in a disc with centre c and radius R.}
}
\value{
A two-dimensional point pattern
}
\description{
Add or remove end point to match a target number of points
}
\examples{
X = endpoint_regrowth(X_thin_end[[90]], num_ep = 40)
X_regrowth = lapply (X_thin_end, function(x) endpoint_regrowth(x,num_ep = 40))
}
