% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENFtrees_class.R
\docType{class}
\name{ENFtrees-class}
\alias{ENFtrees-class}
\title{ENFtrees Class}
\description{
A class "ENFtrees" is represented by two slots
\itemize{
\item Basepoints: A ppp object for the base points
\item Endpoints: A ppp object for the end points
}
}
