% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_cluster_radius.R
\name{Estimate_cluster_radius}
\alias{Estimate_cluster_radius}
\title{Estimates the cluster radius}
\usage{
Estimate_cluster_radius(point_pattern)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
A two dimensional point pattern
}
\description{
Estimates the cluster radius as the maximum distance from an endpoint to the centre of the cluster
}
\examples{
Estimate_cluster_radius(realep[[1]])
}
