% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_list_to_matrix.R
\name{from_list_to_matrix}
\alias{from_list_to_matrix}
\title{List to matrix}
\usage{
from_list_to_matrix(list_)
}
\arguments{
\item{{list_}}{A list of size [N] with each column consisting of m elements
@return A matrix}
}
\description{
Transforms a list of size [N] with m elements to a matrix of the same size.
}
\examples{
list_of_numbers = list(runif(100,0,1),rnorm(100,0,1))
from_list_to_matrix(list_of_numbers)
}
