% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basepoints_ENFtrees.R
\name{get.basepoints_ENFtrees}
\alias{get.basepoints_ENFtrees}
\title{Extract base points from an ENFtree object}
\usage{
get.basepoints_ENFtrees(ENFtree)
}
\arguments{
\item{{ENFtree}}{An "ENFtree" class object with two slots (Basepoints and Endpoints). Each slot is a ppp object}
}
\value{
A ppp object of the base points from an ENFtree object
}
\description{
Extract base points from an ENFtree object and obtain those as ppp objects
}
\examples{
get.basepoints_ENFtrees(X_thin[[1]])
}
\seealso{
\code{\link{get.endpoints_ENFtrees}}
}
