% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basepoints_thinned.R
\name{get.basepoints_thinned}
\alias{get.basepoints_thinned}
\title{Updates the base point patterns.}
\usage{
get.basepoints_thinned(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern for the end point patterns(object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern for the base point patterns(object of class "ppp").
Should contain a mark "Tree".}
}
\value{
The corresponding base point patterns after thinning the end points
}
\description{
After thinning of the end point patterns this function thins the correspondin base point patterns.
This connection is done using the "Tree" mark in the two patterns.
}
\examples{
get.basepoints_thinned(X_thin_end[[1]], realbp[[1]])
}
