% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoints_ENFtress.R
\name{get.endpoints_ENFtrees}
\alias{get.endpoints_ENFtrees}
\title{Extract end points from an ENFtree object}
\usage{
get.endpoints_ENFtrees(ENFtree)
}
\arguments{
\item{{ENFtree}}{An "ENFtree" class object with two slots (Basepoints and Endpoints). Each slot is a ppp object}
}
\value{
A ppp object of the end points from an ENFtree object
}
\description{
Extract end points from an ENFtree object and obtain those as ppp objects
}
\examples{
get.endpoints_ENFtrees(X_thin[[1]])
}
\seealso{
\code{\link{get.basepoints_ENFtrees}}
}
