% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoints_thinned.R
\name{get.endpoints_thinned}
\alias{get.endpoints_thinned}
\title{Updates the end point patterns.}
\usage{
get.endpoints_thinned(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern for the end point patterns(object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern for the base point patterns(object of class "ppp").
Should contain a mark "Tree".}
}
\value{
The corresponding end point patterns after thinning the base points
}
\description{
After thinning of the base point patterns this function thins the correspondin end point patterns.
This connection is done using the "Tree" mark in the two patterns.
}
\examples{
get.basepoints_thinned(realep[[1]], X_thin_base[[1]])
}
