% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpoints_in_tree.R
\name{get.points_in_tree}
\alias{get.points_in_tree}
\title{Returns a ppp object containing the base point and end points of a specific nerve tree.}
\usage{
get.points_in_tree(point_pattern, point_pattern_bp, treeid)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}

\item{{treeid}}{The id of the tree}
}
\value{
A Multi-type point pattern with points in the "treeid" Tree.
}
\description{
Creates a list with $x~coordinates $bbox~bounding box to be used in the Kest_directional function
}
\examples{
get.points_in_tree ( realep[[1]], realbp[[1]], 1)
}
