% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulated_statistics3.R
\name{get.simulated_statistics3}
\alias{get.simulated_statistics3}
\title{Overall centre three-dimenstional L curves from patterns in a list}
\usage{
get.simulated_statistics3(point_pattern_list, nsim = 2500, RMAX = 30)
}
\arguments{
\item{{point_pattern_list}}{A list of size [N] [M] containing three dimensional point patterns ("pp3" objects).}

\item{{nsim}}{Number of simulations. Default is nsim=2500.}

\item{{RMAX}}{Maximum value for the radius parameter}
}
\value{
A matrix where each column is a pooled estimate of the three dimensional centred L function
from patterns that are obtained at random from the list.
}
\description{
Provided a list of size [N] [M], where N=number of samples and M=number of simulations
, at each iteration a random simulated pattern from each sample is selected and the overall centred L
function is estimated. This procedure is repeated nsim times which produces nsim curves.  Square number
weights are used to weights the sample wise summary functions.
}
\examples{
get.simulated_statistics3(simulations)

}
