% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_centres.R
\name{get_cluster_centres}
\alias{get_cluster_centres}
\title{Cluster centres}
\usage{
get_cluster_centres(point_pattern)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
A two dimensional point pattern
}
\description{
Calculates the cluster centre of each cluster
}
\examples{
get_cluster_centres(realep[[1]])
}
