% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoints_centres.R
\name{Get_endpoint.centres}
\alias{Get_endpoint.centres}
\title{Create the point pattern of the end point cluster centres}
\usage{
Get_endpoint.centres(point_pattern)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern for the end point patterns(object of class "ppp").
Should contain a mark "Tree".}
}
\value{
A two-dimensional point pattern for the centre of the end point clusters.
}
\description{
Computes the coordinates of the end points cluster centres and then creates a ppp object where each point represent an end point cluster centre.
}
\examples{
Get_endpoint.centres(realep[[1]])
}
