% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hardcore.R
\name{get_hardcore}
\alias{get_hardcore}
\title{Estimates the hardcore from a point pattern}
\usage{
get_hardcore(point_pattern)
}
\arguments{
\item{{point_pattern}}{Two-dimensional point pattern (object of class "ppp").}
}
\value{
The estimated hardcore parameter h > 0.
}
\description{
Estimates the hardcore parameter h from a point pattern. The estimate is given by h = (N-1) / N * d_min
with N being the number of points in the pattern.
}
\examples{
get_hardcore(rpoisppp(10))
}
