% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_overlap.R
\name{GET_OVERLAP}
\alias{GET_OVERLAP}
\title{Overlap of reactive territories}
\usage{
GET_OVERLAP(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
The overlap of reactive territories. A value between 0 and 1.
}
\description{
Calculates the percentage of overlap of the reactive territories. The Overlap is defined as 1 - ratio, where ratio
is the ratio between the area of the union of the reactive territories divided by the
total area covered by the reactive territories.
}
\examples{
GET_OVERLAP(realep[[1]],realbp[[1]])
}
