% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_union_area.R
\name{GET_UNION_AREA}
\alias{GET_UNION_AREA}
\title{Area of the union of reactive territories}
\usage{
GET_UNION_AREA(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
The areas of the union of the reactive territories (the convex hull of the nerve trees including both end and base points)
}
\description{
Calculates the area of the union of reactive territories
}
\examples{
GET_UNION_AREA(realep[[1]],realbp[[1]])
}
