% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ind_thin.R
\name{ind_thin}
\alias{ind_thin}
\title{Independent random thinning of a point pattern to a fixed number of points}
\usage{
ind_thin(point_pattern, num)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern (object of class "ppp").
Should contain a mark "point_id". The assign_random_marks() function can be used to create such a mark
for a general ppp object.}

\item{{num}}{The desired number of points in the pattern}
}
\value{
Two-dimensional thinned point pattern (object of class "ppp").
}
\description{
Removes points from a pattern until the pattern consists of "num" points
}
\examples{
ind_thin(realep[[1]],10)
}
