% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_to_L_functions.R
\name{K_to_L_cyli}
\alias{K_to_L_cyli}
\title{K to L function transformations}
\usage{
K_to_L_cyli(K, r, epsilon)
}
\arguments{
\item{{K}}{A vector with the directional K functions evaluated at r}

\item{{epsilon}}{The fixed halfwidth of the cylinders (cylindrical functions)
. Central angle for the directed cone ( cone function)}

\item{{r}}{A vector of radiuses to evaluate the summary function.}
}
\value{
A vector with the centred L functions evaluated at r
}
\description{
Transforms directional K functions to directional centred L functions.
}
\examples{
K_to_L_cone(K,r,epsilon)
K_to_L_cyli(K,r,epsilon)
L_to_K_cyli(K,r,epsilon)
}
