% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matclust_with_parents.R
\name{matclust_with_parents_marked}
\alias{matclust_with_parents_marked}
\title{Simulates from matern cluster point process}
\usage{
matclust_with_parents_marked(kappa_par, scale_par, mu_par, window_)
}
\arguments{
\item{{kappa_par}}{Intensity of the Poisson process of cluster centres.}

\item{{scale_par}}{Radius parameter of the clusters.}

\item{{mu_par}}{Mean number of points per cluster (a single positive number)}

\item{{window}}{Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.}
}
\value{
A marked ppp object. This pattern contains two marks "Tree" and "type".
}
\description{
Simulates the daughter and parent process from a matern cluster point process
}
\examples{

parameters_fit =matclust.estK(realep[[66]],
                              startpar=c(kappa=1,scale=1))$clustpar
matclust_with_parents_marked(        parameters_fit[1],
                                    parameters_fit[2],
                                    3,
                                    realep[[66]]$window)
}
