% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_distance.R
\name{minimum_distance}
\alias{minimum_distance}
\title{Closest distance between the points in a point pattern}
\usage{
minimum_distance(point_pattern, minimum = 0)
}
\arguments{
\item{{point_pattern}}{Two-dimensional point pattern (object of class "ppp").}

\item{{minimum}}{The percentile of the distance to consider. For example, minimum=0 corresponds to the min distance, minimum =0.1 is the 10% percentile and so on.
Default value is minimum = 0 which returns the closest distance}
}
\value{
A vector with the minimum distances
}
\description{
For each point in the point pattern it computes the distances to the other points and returns the "minimum" percentile of those distances. If
minimum = 0 the minimum distance is returned. This function is mainly used to create distance based marks and thin the patterns according to the values of those marks.
}
\examples{
minimum_distance(rpoispp(10))
}
