% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_centred_L3est.R
\name{Overall_centred_L3est}
\alias{Overall_centred_L3est}
\title{Overall three-dimensional L(r) - r function}
\usage{
Overall_centred_L3est(list_of_patterns, RMAX = 30)
}
\arguments{
\item{{list_of_patterns}}{A list containing three-dimensional point patterns (objects of class "pp3").}

\item{{RMAX}}{Maximum value for the radiuses. The function will be evaluated  in (0, RMAX ) at 100 different radius values.}
}
\value{
A dataframe with columns
\itemize{
\item $r:  radiuses where the summary function is evaluated and
\item $L:  a vector containing the values of the summary statistc for each r}
}
\description{
Computes the overall  L(r) - r function from a list of pp3 objects.
The Ripley's K(r) function for each sample is weighted using square number points weights.
}
\examples{
Overall_centred_L3est(realep3D)
}
