% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_centred_Lcyl.R
\name{Overall_centred_Lcyl}
\alias{Overall_centred_Lcyl}
\title{Overall cylindrical and centred L function}
\usage{
Overall_centred_Lcyl(
  list_of_xsim,
  rad_p = 15,
  rad_z = 15,
  r = seq(0, 30, length = 200)
)
}
\arguments{
\item{{list_of_xsims}}{A list containing xsim objects( a dataframe with collumns $x~coordinates and $bbox~bounding box)}

\item{{rad_p}}{The fixed halfwidth of the cylinders in the x and y directions}

\item{{rad_z}}{The fixed halfwidth of the cylinders in the z direction.}

\item{{r}}{A vector of radiuses to evaluate the summary function.}
}
\value{
A dataframe with columns :
\itemize{
\item $r : radiuses where the summary functions are evaluated
\item $L_x : a vector containing the values of the summary statistc directed towards the x direction for each r
\item $L_y : a vector containing the values of the summary statistc directed towards the y direction for each r and
\item $L_z : a vector containing the values of the summary statistc directed towards the z direction for each r
}
}
\description{
Computes the overall cylindical and centred L function  from a list of xsim objects directed towards the three main axes
The cylindrical K(r) function for each sample is weighted using square number points weights.
}
\examples{
Overall_centred_Lcyl(list_of_xsim)
}
