% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_centred_Lest.R
\name{Overall_centred_Lest}
\alias{Overall_centred_Lest}
\title{Overall two-dimensional L(r) - r function}
\usage{
Overall_centred_Lest(list_of_patterns, RMAX = 30)
}
\arguments{
\item{{list_of_patterns}}{A list containing two-dimensional point patterns (objects of class "ppp").}
}
\value{
A dataframe with columns
\itemize{
\item $r:  radiuses where the summary function is evaluated and
\item $L:  a vector containing the values of the summary statistc for each r}
}
\description{
Computes the overall  L(r) - r function from a list of ppp objects.
The Ripley's K(r) function for each sample is weighted using square number points weights.
}
\examples{
Overall_centred_Lest(realep[1:4])
}
