% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_statistics.R
\name{overall_statistics}
\alias{overall_statistics}
\title{Calculates non-spatial statistics}
\usage{
overall_statistics(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{A list of two-dimensional point patterns for the end points(object of class "ppp").
Should have a mark "Tree".}

\item{{point_pattern_bp}}{A list of two-dimensional point patterns for the base points(object of class "ppp").
Should have a mark "Tree".}
}
\value{
A dataframe with columns
\itemize{
\item $num_ep : Number of end points
\item $area_per_tree : Average area of reactive territory
\item $trees_per_sample :  Number of nerve trees
\item $points_per_cluster : Number of end points per nerve tree
\item $orphan_trees : Number of clusters with one end point
}
}
\description{
Calculates non-spatial statistics from the patterns.
}
\examples{
overall_statistics(point_pattern = realep,
                   point_pattern_bp = realbp)
}
