% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_thinned_process.R
\name{plot_thinned_process}
\alias{plot_thinned_process}
\title{Plots the thinned patterns}
\usage{
plot_thinned_process(
  ppp_real,
  ppp_bp_real,
  ppp_thinned,
  ppp_bp_thinned,
  plot.basepoints = TRUE,
  plot.legend = TRUE,
  main = "",
  connection = TRUE
)
}
\arguments{
\item{{ppp_real}}{Two-dimensional point pattern for the end point patterns(object of class "ppp") prior to thinning.}

\item{{ppp_bp_real}}{Two-dimensional  point pattern for the base point patterns(object of class "ppp") prior to thinning.}

\item{{ppp_thinned}}{Two-dimensional point pattern for the end point patterns(object of class "ppp") after thinning.}

\item{{ppp_bp_thinned}}{Two-dimensional  point pattern for the base point patterns(object of class "ppp") after thinning.}

\item{{plot.basepoints}}{Graphical parameter that controls whether to plot the base point patterns}

\item{{plot.legend}}{Graphical parameter that controls whether to add a legend.}

\item{{main}}{Graphical parameter that controls the main title.}

\item{{connection}}{If TRUE it adds the segment between base and end points when plot.basepoint=TRUE}
}
\value{
Creates a plot
}
\description{
Function that plots the thinned patterns and compare it with the original pattern prior to thinning.
Used to visualize the thinned patterns and compare them with the original patterns before applying any thinning.
}
\examples{
plot_thinned_process(realep[[1]],
realbp[[1]],
X_thin_end[[1]],
X_thin_base[[1]],
plot.basepoints = FALSE,
plot.legend = FALSE)
}
