% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_retain_points.R
\name{probability.retain_points}
\alias{probability.retain_points}
\title{Probability function based on closest distances and cluster sizes}
\usage{
probability.retain_points(
  point_pattern,
  weight_ep = 0.045,
  scheme = "dependent"
)
}
\arguments{
\item{{point_pattern}}{Two-dimensional point pattern for the end points(object of class "ppp").}

\item{{weight_ep}}{the scale parameter w of the probability function.}

\item{{scheme}}{The thinning scheme for the end points. Either "dependent" or "random". Default is "dependent"}
}
\value{
A vector of probabilities
}
\description{
A certain probability function for thinning the end points
 given by probability = exp (- ( w* min_distance)^2) if size !=1 oterwise probability=1
, where size is the cluster size.
}
\examples{
probability.retain_points(X_thin_end[[63]] )
}
