% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROBABILITY_THINNING_.R
\name{PROBABILITY_THINNING_}
\alias{PROBABILITY_THINNING_}
\title{Probability function based on closest distances}
\usage{
PROBABILITY_THINNING_(
  point_pattern,
  point_pattern_bp,
  weight_bp = 0.013,
  weight_ep = 0.045,
  minimum = 0,
  type = "BP",
  use_square = TRUE
)
}
\arguments{
\item{{point_pattern}}{Two-dimensional point pattern for the end points(object of class "ppp").}

\item{{point_pattern_bp}}{Two-dimensional point pattern for the base points(object of class "ppp").}

\item{{weight_bp}}{scale parameter if type = "BP". Those parameters need to be tuned.}

\item{{weight_ep}}{scale parameter if type = "EP". Those parameters need to be tuned.}

\item{{minimim}}{The percentile of the distance to consider. For example, minimum=0 corresponds to the min distance, minimum =0.1 is the 10% percentile and so on.
Default value is minimum = 0 which returns the closest distance}

\item{{type}}{This argument define which pattern should be thinned
\itemize{
\item "BP" : Base points
\item "EP" : End points
}}

\item{{use_square}}{A parameter that defines whether the probability function will be gaussian of exponential.
Default = TRUE which gives gaussian like probability}
}
\value{
A vector of probabilities
}
\description{
A certain probability function for thinning the points (either base or end points). The functions is
given by probability = exp (- ( w* min_distance)^2)
}
\examples{
PROBABILITY_THINNING_(realep[[1]],realbp[[1]],weight_bp = 0.05)
}
