% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_samples.R
\name{Random.samples}
\alias{Random.samples}
\title{Chooses samples at random}
\usage{
Random.samples(list_of_ENFtrees, nsim = 2500)
}
\arguments{
\item{{list_of_ENFtrees}}{list of size [N][M]}

\item{{nsim}}{Number of indices to simulate. Default value is nsim = 2500.}
}
\value{
A (N) x (nsim) matrix containing random indices between 1 and M.
}
\description{
From a list of size [N][M] create a matrix of size (N) x (nsim) containing random indices between 1 and M.
}
\examples{
Random.samples(simulations)
}
