% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_tree_thinning.R
\name{random_tree_thinning}
\alias{random_tree_thinning}
\title{Function for random tree-wise thinning}
\usage{
random_tree_thinning(ppp, prob)
}
\arguments{
\item{{point_pattern}}{A two-dimensional point pattern(object of class "ppp" in spatstat).
Should have a mark "Tree".}

\item{{prob}}{A real number giving the fixed probability}
}
\value{
A thinned point pattern
}
\description{
This function will thin trees(all points in the tree are removed) according to a fixed probability prob.
}
\examples{
random_tree_thinning( marked_point_pattern , prob)
}
