% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_points_hardcore.R
\name{remove.points_hardcore}
\alias{remove.points_hardcore}
\title{Apply the hardcore radius to a pattern}
\usage{
remove.points_hardcore(point_pattern, h)
}
\arguments{
\item{{point_pattern}}{Two-dimensional point pattern (object of class "ppp").}

\item{{h}}{The hardcore parameter h > 0.}
}
\value{
A two-dimensional points pattern (object of class "ppp") with minimum distance between the points larger than the hardcore parameter h.
}
\description{
This function remove all points in a point pattern that do not respect the hardcore i.e.
If two points are closer than distance h apart one of them is removed.
}
\examples{
remove.points_hardcore(rpoispp(10),0.2)
}
