% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_in_convex_hull.R
\name{Simulate_in_convex_hull}
\alias{Simulate_in_convex_hull}
\title{Suggest point in reactive territory}
\usage{
Simulate_in_convex_hull(point_pattern, point_pattern_bp, Treeid, hc = 0.62)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}

\item{{Treeid}}{The Id of the tree}

\item{{hc}}{The hardcore parameter}
}
\value{
A two dimensional point pattern with one point
}
\description{
Simulate a uniformly random point in reactive territory
}
\examples{
Simulate_in_convex_hull(realep[[1]],realbp[[1]],1)
}
