% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_thinned_basepointpatterns.R
\name{simulate_thinned_basepointpatterns}
\alias{simulate_thinned_basepointpatterns}
\title{Simulates thinned patterns}
\usage{
simulate_thinned_basepointpatterns(
  sample_num = 1,
  point_pattern_target = realepmild,
  point_pattern_bp_target = realbpMILD,
  normal_patterns = realep,
  normal_bp_patterns = realbp,
  weight_bp = 0.1,
  scheme = "dependent",
  simulate_healthy = FALSE,
  simulation.params = c(0.0003409953, 22.73218, 3),
  simulation.size = 40,
  hardcorestep = TRUE
)
}
\arguments{
\item{{sample_num}}{The sample number}

\item{{normal_patterns}}{A  list of two-dimensional point pattern for the end points to be thinned(object of class "ppp").
Should have a mark "Tree".}

\item{{normal_bp_patterns}}{A list of two-dimensional point patterns for the base points to be thinned(object of class "ppp").
Should have a mark "Tree".}

\item{{point_pattern_target}}{A list of two-dimensional point patterns for the end points
(object of class "ppp"). The number of point in this pattern give the target end point in the thinned pattern.
Should have a mark "Tree".}

\item{{point_pattern_bp_target}}{A list of two-dimensional point patterns for the base points to be thinned(object of class "ppp").
The number of points in this pattern give the target base point in the thinned pattern.
Should have a mark "Tree".}

\item{{weight_bp}}{A scaling parameter for the probability function}

\item{{simulate_healthy}}{A variable deciding whether to simulate or use the empirical healthy data as a starting point}

\item{{simulation.params}}{If simulate_healthy=TRUE. The healthy data are simulated from a Matern Cluster process
with simulation.params (kappa,scale,mu).}

\item{{hardcorestep}}{Whetther to use the hardcore step or not}

\item{{simulation.size}}{Defines how many healthy samples to simulate from the Matern Cluster process}
}
\value{
A (list) of ENFtrees objects.
}
\description{
Thins given patterns according to a parametric thinning model. First two steps of the model in the paper.
}
\examples{
simulate_thinned_basepointpatterns()
simulate_thinned_basepointpatterns(normal_bp_patterns = realbp[[66]],normal_patterns = realep[[66]])
}
