% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_thinned_patterns.R
\name{simulate_thinned_patterns}
\alias{simulate_thinned_patterns}
\title{Simulates thinned patterns}
\usage{
simulate_thinned_patterns(
  sample_num,
  point_pattern_target = realepmild,
  point_pattern_bp_target = realbpMILD,
  normal_patterns = realep,
  normal_bp_patterns = realbp,
  weight_bp = 0.3,
  scheme = "dependent",
  regrowth = TRUE,
  R = 20
)
}
\arguments{
\item{{sample_num}}{The sample number}

\item{{point_pattern}}{A list of two-dimensional point patterns for the end points to be thinned(object of class "ppp").
Should have a mark "Tree".}

\item{{point_pattern_bp}}{A list of two-dimensional point patterns for the base points to be thinned(object of class "ppp").
Should have a mark "Tree".}

\item{{point_pattern_target}}{A list of two-dimensional point patterns for the end points
(object of class "ppp"). The number of point in this pattern give the target end point in the thinned pattern.
Should have a mark "Tree".}

\item{{point_pattern_bp_targe}}{A list of two-dimensional point patterns for the base points to be thinned(object of class "ppp").
The number of points in this pattern give the target base point in the thinned pattern.
Should have a mark "Tree".}

\item{{weight_bp}}{A scaling parameter for the probability function}

\item{{scheme}}{A string determining the type of thinning to be applied. Either "dependent" or "random".}

\item{{regrowth}}{A boolean variable. If it is true a regrowth model(add or remove end points to match the target points) will be used}

\item{{R}}{The radius parameter of the regrowth model when we should add end points}
}
\value{
A list of ENFtrees objects.
}
\description{
Thins given patterns to a target number of base and end points. We perform the three step thinning procedure described in the paper.
This is done is performed with thin_to_sample() and endpoint_regrowth() functions.
}
\examples{
simulate_thinned_patterns(sample_num = 1,scheme="dependent")
}
