% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_uniformly_in_disc.R
\name{simulate_uniformly_in_disc}
\alias{simulate_uniformly_in_disc}
\title{Uniform points in disc}
\usage{
simulate_uniformly_in_disc(n, centre, r)
}
\arguments{
\item{{n}}{number of points to return.}

\item{{centre}}{the $x~coordinates and $y~coordinates for the disc centre}

\item{{r}}{the radius of the disc}
}
\value{
a dataframe with the $x~coordinates and $y~coordinates of the simulated points
}
\description{
Creates a dataframe with the $X~coordinates and $Y~coordinates for the simulated points
}
\details{
Simulates uniform points in the disc with centre and Radius r
}
\examples{
simulate_uniformly_in_disc(1000,c(0,0),3)

}
