% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCC_model.R
\name{simulateUCC}
\alias{simulateUCC}
\title{UCC model}
\usage{
simulateUCC(Basepoints, Group, type = "UCC", parameters = 0, ID = 99)
}
\arguments{
\item{{Basepoints}}{Dataframe for the base points patterns}

\item{{Group}}{Normal of Mild}

\item{{type}}{"UCC" or "NOC"}

\item{{params}}{Data frame with the estimated parameters}

\item{{ID}}{ID of subject}
}
\value{
A dataframe
}
\description{
imulates EP codnitioned on BP using the NOC/UCC models
}
\examples{
simulateUCC(realbpdf[[1]],"Normal",type="NOC",parameters = params_subject,ID=SID[1])
simulateUCC(get.basepoints(Foot.1097.1.pgp1),"Normal")
simulateUCC(realbpdf[[1]],"Normal",parameters = Parameters_normal,ID=SID[1])
}
