% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_to_sample.R
\name{thin_to_sample}
\alias{thin_to_sample}
\title{Thins patterns to specific number of points}
\usage{
thin_to_sample(
  ppp_ep,
  ppp_bp,
  num_ep,
  num_bp,
  epsilon = 0,
  scheme = "random",
  weight_bp = 0.2,
  weight_ep = 0.045
)
}
\arguments{
\item{{ppp_ep}}{Two-dimensional marked point pattern for the end point patterns(object of class "ppp").
Should contain a mark "Tree".}

\item{{ppp_bp}}{Two-dimensional marked point pattern for the base point patterns(object of class "ppp").
Should contain a mark "Tree".}

\item{{num_ep}}{Target number of end points. Notice that if num_ep is large, only the base points will be thinned}

\item{{num_bp}}{Target number of base points. Notice that if num_bp is large, only the end points will be thinned}

\item{{weight_bp}}{scale parameter in the PROBABILITY_THINNING() function}

\item{{weight_ep}}{scale parameter in the PROBABILITY_THINNING() function}

\item{{scheme}}{A string determining the type of thinning to be applied. Either "dependent" or "random".}

\item{{epsilon}}{epsilon error allowed, i.e. thin to num_ep+epsilon  points instead}
}
\value{
An ENFtrees object for the thinned pattern.
}
\description{
Applies a certain thinning strategy to thin patterns to a target number of points.
}
\examples{
thin_to_sample(realep[[1]],
realbp[[1]],
num_bp = 40,
epsilon =0,
num_ep = 10000 ,
weight_bp = 0.005,
weight_ep = 0.008,
scheme="dependent")
}
