% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_area.R
\name{total_area}
\alias{total_area}
\title{Computes the total area of the skin covered by the ENFs}
\usage{
total_area(point_pattern, point_pattern_bp)
}
\arguments{
\item{{point_pattern}}{Two-dimensional marked point pattern of the end points (object of class "ppp").
Should contain a mark "Tree".}

\item{{point_pattern_bp}}{Two-dimensional marked point pattern of the base points (object of class "ppp").
Should contain a mark "Tree".}
}
\value{
A data frame with columns Total_area (total area of reactive territories in the pattern),
Trees( Number of nerve trees in the patterns) and area_per_tree( The average reactive territory area per tree)
}
\description{
Computes the total and average area of reactive territories given two point patterns( base and end point patterns)
}
\examples{
total_area(realep[[1]],realbp[[1]])
}
