% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_regrowth.R
\name{endpoint_regrowth}
\alias{endpoint_regrowth}
\title{Regrowth model}
\usage{
endpoint_regrowth(point_pattern, num_ep, hc = 1, R = 20, weight_ep = 0.045)
}
\arguments{
\item{{point_pattern}}{A two-dimensional point pattern (object of class "ppp").
Should have a mark "Tree.}

\item{{num_ep}}{Target number of end points}

\item{{hc}}{Hardcore parameter}

\item{{R}}{Radius parameter in the regrowth model when adding points}

\item{{weight_ep}}{Scaling parameter in the probabiity function.}
}
\value{
A two-dimensional point pattern
}
\description{
Add or remove end point to match a target number of points
}
\examples{
X = endpoint_regrowth(X_thin_end[[90]], num_ep = 40)
X_regrowth = lapply (X_thin_end, function(x) endpoint_regrowth(x,num_ep = 40))
}
